# server/docs/swagger.yaml
basePath: /v1
definitions:
...
  response.Response:
    description: Structure for a standard API response
    properties:
      data: {}
      error:
        example: any error
        type: string
      status:
        example: success/error
        type: string
    type: object
info:
  contact:
    name: Evdokimov Igor
    url: https://t.me/epelptic
  description: API for potatorate site
  title: Film-catalog API
  version: 1.0.0
paths:
...
  /films:
    get:
      description: Возвращает список фильмов с возможностью фильтрации по жанрам,
        актерам, продюсеру, рейтингу, дате выпуска, длительности и сортировке
      parameters:
      - collectionFormat: csv
        description: Список ID жанров
        in: query
        items:
          type: integer
        name: genre_ids
        type: array
      - collectionFormat: csv
        description: Список ID актеров
        in: query
        items:
          type: integer
        name: actor_ids
        type: array
      - description: Продюсер
        in: query
        name: producer
        type: string
      - description: Минимальный рейтинг
        in: query
        name: min_rating
        type: number
      - description: Максимальный рейтинг
        in: query
        name: max_rating
        type: number
      - description: 'Минимальная дата выпуска (формат: 2006-01-02)'
        in: query
        name: min_date
        type: string
      - description: 'Максимальная дата выпуска (формат: 2006-01-02)'
        in: query
        name: max_date
        type: string
      - description: 'Минимальная длительность (пример: 2h30m, 90m)'
        in: query
        name: min_duration
        type: string
      - description: 'Максимальная длительность (пример: 2h30m, 90m)'
        in: query
        name: max_duration
        type: string
      - description: Поле для сортировки (rating, release_date, runtime)
        in: query
        name: sort_by
        type: string
      - description: Порядок сортировки (asc, desc)
        in: query
        name: order
        type: string
      - description: Номер страницы
        in: query
        name: page
        type: integer
      - description: Размер страницы
        in: query
        name: page_size
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.Response'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Получить список фильмов с фильтрацией и пагинацией
      tags:
      - film
    post:
      consumes:
      - multipart/form-data
      description: Создает новый фильм с указанными параметрами
      parameters:
      - description: Данные фильма в формате JSON
        in: formData
        name: data
        required: true
        type: string
      - description: Постер фильма
        in: formData
        name: poster
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Создать новый фильм
      tags:
      - film
  /films/{id}:
    delete:
      description: Удаляет фильм по его ID
      parameters:
      - description: ID фильма
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Удалить фильм по ID
      tags:
      - film
    get:
      description: Возвращает информацию о фильме по его ID
      parameters:
      - description: ID фильма
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Получить фильм по ID
      tags:
      - film
    put:
      consumes:
      - multipart/form-data
      description: Обновляет данные фильма, включая постер
      parameters:
      - description: ID фильма
        in: path
        name: id
        required: true
        type: string
      - description: Данные фильма в формате JSON
        in: formData
        name: data
        required: true
        type: string
      - description: Постер фильма
        in: formData
        name: poster
        type: file
      - description: Удалить постер
        in: formData
        name: remove_poster
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Обновить информацию о фильме
      tags:
      - film
  /films/search:
    get:
      description: Возвращает список фильмов, соответствующих поисковому запросу
      parameters:
      - description: Поисковый запрос
        in: query
        name: query
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.Response'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Поиск фильмов по запросу
      tags:
      - film
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"